feat: build SAP Commerce bundle from external directory
When building SAP Commerce we try to minize the volume size as much as
possible for the generated containers. To achieve this the base
container now incorporates the unpacked fails from a `target/`
directory.

This `target/` dir will only contains the necessary files to run the
application and the installation scripts. All data contained in the
bundle have been removed so the image size is kept at a minimum.