feat: build SAP Commerce from external directory
Prepares Hybris bundle outside the container build. This has the
advantage of keeping the docker image size to a minimum.

With this commit we reverted back the old docker build files in the
project home directory and moved the new externalised build to `ybase`.
In the future we will introduce a two-phased build with a base image
plus a final one with code customizations.

Still lacking documentation, will introduce in subsequent commits.

Fixes #12